Fix imports
Fix imports for modified js structure
