+++ src/auth/src/main/resources/static/js/listeners/listeners-archetypes/generic-dom-listener.js
@@ -1,7 +1,8 @@
 "use strict";
 
-import { DOM_DATA_GETTERS, parseDomData } from "../utils/parser/dom-parser.js";
-import validateFields from "../utils/validators/validate-fields.js";
+import DOM_DATA_GETTERS from "../../utils/parser/dom/dom-getters.js";
+import parseDomData from "../../utils/parser/dom/dom-parser.js";
+import validateFields from "../../utils/validators/fields/validate-fields.js"
 
 const genericDomListener = (names, service) => {
 	const fields = parseDomData(DOM_DATA_GETTERS.name, ...names);
