+++ src/auth/src/main/resources/static/js/listeners/register.js
@@ -1,7 +1,7 @@
 "use strict";
 
 import { userAuthService } from "../services/services.js";
-import { DOM_DATA_GETTERS } from "../utils/parser/dom-parser.js";
+import DOM_DATA_GETTERS from "../../utils/parser/dom/dom-getters.js";
 import genericDomListener from "./listeners-archetypes/generic-dom-listener.js";
 
 const names = ["username", "email", "passwd"];
