+++ src/auth/src/main/resources/static/js/services/callbacks/auth-success.js
@@ -1,7 +1,7 @@
 "use strict";
 
 import { getJwt } from "../../utils/fetch/jwt/jwt-fetch-methods.js";
-import ROUTES from "../../data/routes.js";
+import ROUTES from "../../data/routes/routes.js";
 
 const authSuccess = (token) => {
 	const REDIRECT_ROUTE = ROUTES.user;
