+++ src/auth/src/main/resources/static/js/services/callbacks/passwd-reset-success.js
@@ -1,7 +1,7 @@
 "use strict";
 
 import { get } from "../../utils/fetch/methods/fetch-methods.js";
-import ROUTES from "../../data/routes.js";
+import ROUTES from "../../data/routes/routes.js";
 
 const passwdResetSuccess = (...args) => {
 	const REDIRECT_ROUTE = ROUTES.login;
