+++ src/auth/src/main/resources/static/js/utils/validators/validate/generic/validate.js
