+++ src/auth/src/main/resources/static/js/utils/validators/validate/implementations/username-validator.js
@@ -1,6 +1,7 @@
 "use string";
 
-import { validate, VALIDATOR_EXPRESSIONS } from "../generic/validate.js";
+import VALIDATOR_EXPRESSIONS from "../../data/validator-expression.js"
+import validate from "../generic/validate.js";
 
 const validateUsername = (username) => validate(username, VALIDATOR_EXPRESSIONS.username);
