--- src/db/README.md
@@ -1,3 +1,0 @@
-# SQL directory
-
-Directory for SQL scripts
